+++ app/Controller/UmbrellaOrganisationsController.php
@@ -138,12 +138,14 @@ class UmbrellaOrganisationsController extends AppController
     public function delete($id) {
         $this->isAuthorizedAs(['Administrator', 'Account manager', 'School manager', 'School management']);
 
-        $result = $this->UmbrellaOrganisationsService->deleteOrganisation($id);
+        if($this->request->is('delete')) {
+            $result = $this->UmbrellaOrganisationsService->deleteOrganisation($id);
 
-        $this->formResponse(
-            $result ? true : false,
-            []
-        );
+            $this->formResponse(
+                $result ? true : false,
+                []
+            );
+        }
     }
 
     public function add() {
